更新关于界面，更新莹石sdk

+++ BabyQ/Video_libs/Headers/EZProbeDeviceInfo.h
@@ -45,7 +45,7 @@
 /**
  *  N1，R1，A1等设备关联的设备数
  */
-@property (nonatomic) int relatedDeviceCount;
+@property (nonatomic) NSInteger relatedDeviceCount;
 /**
  *  能力集
  */
